Change order amount to float and adjust schema

+++ server/app/models/order_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, BigInteger, String, TIMESTAMP, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, TIMESTAMP, ForeignKey, Float
 from sqlalchemy.orm import relationship
 from app.config.database import Base
 
@@ -10,7 +10,7 @@ class Order(Base):
     client_id = Column(Integer, ForeignKey("clients.id"), nullable=False)
     created_date = Column(TIMESTAMP(timezone=False), nullable=False)
     status = Column(String(255), nullable=False)
-    amount = Column(BigInteger, nullable=False)
+    amount = Column(Float, nullable=False)
 
     client = relationship("Client", back_populates="orders")
     order_items = relationship("OrderItem", back_populates="order", cascade="all, delete-orphan")
